add combineReducer && change files structure
