+++ src/components/partials/catalog/catalog.jsx
@@ -5,7 +5,7 @@ import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import {DEFAULT_FILTER, MOVIES_COUNTER_INITIAL} from "@constants";
 
-import {setActiveFilter, increaseMoviesCounter} from "@store/action-creators";
+import {setActiveFilter, increaseMoviesCounter} from "@store/movies-data/action-creators";
 import {getActiveFilter, getGenres, getMoviesCounter} from "@store/selectors";
 
 import CatalogCard from "@partials/catalog-card/catalog-card";
