+++ src/store/common/action-types.js
@@ -1,0 +1,4 @@
+export const ActionTypes = {
+  SET_FETCHING_MOVIES: `SET_FETCHING_MOVIES`,
+  SET_FETCHING_PROMO: `SET_FETCHING_PROMO`,
+};
