+++ src/store/movies-data/action-creators.js
@@ -1,4 +1,4 @@
-import {ActionTypes} from "./action-types";
+import {ActionTypes} from "@store/movies-data/action-types";
 import {MOVIES_COUNTER_STEP} from "@constants";
 
 export const setMovies = (movies) => {
@@ -26,26 +26,5 @@ export const increaseMoviesCounter = () => {
   return {
     type: ActionTypes.INCREASE_MOVIES_COUNTER,
     payload: MOVIES_COUNTER_STEP
-  };
-};
-
-export const setIsAuthorized = (isAuthorized) => {
-  return {
-    type: ActionTypes.SET_IS_AUTHORIZED,
-    payload: isAuthorized
-  };
-};
-
-export const setFetchingMovies = (isLoaded) => {
-  return {
-    type: ActionTypes.SET_FETCHING_MOVIES,
-    payload: isLoaded
-  };
-};
-
-export const setFetchingPromo = (isLoaded) => {
-  return {
-    type: ActionTypes.SET_FETCHING_PROMO,
-    payload: isLoaded
   };
 };
