+++ src/store/movies-data/action-types.js
@@ -3,7 +3,4 @@ export const ActionTypes = {
   SET_PROMO_MOVIE: `SET_PROMO_MOVIE`,
   SET_ACTIVE_FILTER: `SET_ACTIVE_FILTER`,
   INCREASE_MOVIES_COUNTER: `INCREASE_MOVIES_COUNTER`,
-  SET_IS_AUTHORIZED: `SET_IS_AUTHORIZED`,
-  SET_FETCHING_MOVIES: `SET_FETCHING_MOVIES`,
-  SET_FETCHING_PROMO: `SET_FETCHING_PROMO`,
 };
