+++ src/store/movies-data/movies-data.js
@@ -1,18 +1,15 @@
-import {ActionTypes} from "./action-types";
-import {getAdaptedMovies, getAdaptedMovie} from "@api/utils";
+import {ActionTypes} from "@store/movies-data/action-types";
+import {getAdaptedMovie, getAdaptedMovies} from "@api/utils";
 import {DEFAULT_FILTER, MOVIES_COUNTER_INITIAL} from "@constants";
 
-export const initialState = {
+const initialState = {
   movies: [],
   promoMovie: {},
   activeFilter: DEFAULT_FILTER,
   moviesCounter: MOVIES_COUNTER_INITIAL,
-  isAuthorized: false,
-  isLoadingMovies: false,
-  isLoadingPromo: false,
 };
 
-export const rootReducer = (state = initialState, action) => {
+export const reducer = (state = initialState, action) => {
   switch (action.type) {
     case ActionTypes.SET_MOVIES:
       return Object.assign({}, state, {
@@ -32,20 +29,6 @@ export const rootReducer = (state = initialState, action) => {
     case ActionTypes.INCREASE_MOVIES_COUNTER:
       return Object.assign({}, state, {
         moviesCounter: state.moviesCounter + action.payload,
-      });
-
-    case ActionTypes.SET_IS_AUTHORIZED:
-      return Object.assign({}, state, {
-        isAuthorized: action.payload,
-      });
-
-    case ActionTypes.SET_FETCHING_MOVIES:
-      return Object.assign({}, state, {
-        isLoadingMovies: action.payload,
-      });
-    case ActionTypes.SET_FETCHING_PROMO:
-      return Object.assign({}, state, {
-        isLoadingPromo: action.payload,
       });
   }
