+++ src/store/selectors.js
@@ -1,16 +1,16 @@
 import {createSelector} from "reselect";
 import {DEFAULT_FILTER} from "@constants";
 
-export const getMovies = (state) => state.movies;
-export const getActiveFilter = (state) => state.activeFilter;
-export const getPromoMovie = (state) => state.promoMovie;
-export const getMoviesCounter = (state) => state.moviesCounter;
-export const getMovieById = (state, movieId) => state.movies.find(({id}) => id === parseInt(movieId, 10));
-export const getFetchingMovies = (state) => state.isLoadingMovies;
-export const getFetchingPromo = (state) => state.isLoadingPromo;
+export const getMovies = (state) => state.moviesData.movies;
+export const getActiveFilter = (state) => state.moviesData.activeFilter;
+export const getPromoMovie = (state) => state.moviesData.promoMovie;
+export const getMoviesCounter = (state) => state.moviesData.moviesCounter;
+export const getMovieById = (state, movieId) => state.moviesData.movies.find(({id}) => id === parseInt(movieId, 10));
+export const getFetchingMovies = (state) => state.common.isLoadingMovies;
+export const getFetchingPromo = (state) => state.common.isLoadingPromo;
 
 export const getRelatedMovies = (state, currentMovie) =>
-  state.movies.filter(({id, genre}) => id !== currentMovie.id && genre === currentMovie.genre);
+  state.moviesData.movies.filter(({id, genre}) => id !== currentMovie.id && genre === currentMovie.genre);
 
 export const getGenres = createSelector(
     [getMovies],
