+++ src/store/user/action-creators.js
@@ -1,0 +1,8 @@
+import {ActionTypes} from "@store/user/action-types";
+
+export const setIsAuthorized = (isAuthorized) => {
+  return {
+    type: ActionTypes.SET_IS_AUTHORIZED,
+    payload: isAuthorized
+  };
+};
