+++ src/store/user/action-types.js
@@ -1,0 +1,3 @@
+export const ActionTypes = {
+  SET_IS_AUTHORIZED: `SET_IS_AUTHORIZED`,
+};
